Feature/abandoned cart tracking (#4)
* fix(spree_chimpy, order_upserter, orders): Fix issue with no autoloading in Rails 5 breaking gem

* fix(spree_chimpy): Fix lazyload files since Rails.application is not loaded yet

* fix(lib/spree/chimpy/interface/products.rb): Fix MailChimp tracking integrations

Modify the self.variant_hash method of the Products interface
-- Only include the :image_url if a
vairant image is present. This is an optional key, but causes an error in the MailChimp API if set
to nil. It should be excluded if it's not defined on the object.

Fix bug in MailChimp Campaign tracking
-- Store MailChimp params in session hash until they can be attached to an order

* feat(carts): Add support for tracking Carts in MailChimp
Add integration with the Carts endpoint in the MailChimp
API to allow the use of Abandoned Cart automations

feat(cart tracking): Add support for tracking Carts in MailChimp

-- Add support for tracking carts in MailChimp to allow the use of Abandoned Cart campaigns.

* fix(spree_chimpy.rb): Add missing require calls for Carts interface

Add missing call to require rb file for the Carts interface

* fix(carts/orders): Fix sync of carts/orders

Fix the sync process for Carts and Orders so both data elements sync correctly with MailChimp.
--
Add before_action calls in the `checkouts_controller.rb` so carts update appropriately at all
stages
-- change orders_controller to call `order.notify_mail_chimp` as `after_action` (this
ensures that during checkout completion, an order is synced AFTER the checkout is complete rather
than after a "complete order" request is submitted